third_party/blink/web_tests/platform/generic/fullscreen/full-screen-element-stack-expected.txt

Test for W3C Fullscreen element stack.

To test manually, click the "Go full screen" button - the page should enter full screen mode.

Go full screen (one)
Go full screen (two)
Exit full screen (two)
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitFullscreenElement.id == 'one') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitFullscreenElement.id == 'two') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitFullscreenElement.id == 'one') OK
END OF TEST
